getConfirmations should be declared as int

+++ src/main/java/net/helix/pendulum/controllers/TransactionViewModel.java
@@ -558,7 +558,7 @@ public class TransactionViewModel {
     }
     public void setRoundIndex(long roundIndex) { transaction.roundIndex = roundIndex; }
 
-    public long getConfirmations() {
+    public int getConfirmations() {
         return transaction.confirmations;
     }
     public void setConfirmations(int confirmations) { transaction.confirmations = confirmations; }
